Split the code into its respective files, lowered the learning rate of the scheduler and the optimiser to 1e-4 because the loss was diverging. Updated READMEto contain description of the network.
