+++ recognition/Siamese_45330212/predict.py
@@ -1,3 +1,24 @@
 # Shows example usage of my trained model. Print out any results and / or provide visualisations where applicable
+import torchvision
+import torch.utils.data as utils
+from torchvision import datasets
+import torchvision.transforms as transforms
+from torch.utils.data import DataLoader,Dataset
+from torch.autograd import Variable
+import matplotlib.pyplot as plt
+import torchvision.utils
+import numpy as np
+from torch.optim import lr_scheduler
+import os
+from PIL import Image
 import torch
+from torch.autograd import Variable
 import torch.nn as nn
+from torch import optim
+import torch.nn.functional as F
+import pandas as pd
+import random
+import time
+from modules import *
+from dataset import *
+from train import *
